requirements.txt
asgiref==3.7.2
boto3==1.34.4
botocore==1.34.4
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
jmespath==1.0.1
kafka==1.3.5
kafka-python==2.0.2
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
